fix: product price became float instead of uint

+++ internal/entity/product.go
@@ -1,7 +1,7 @@
 package entity
 
 type Product struct {
-	ID          uint   `json:"id" gorm:"primaryKey"`
-	Description string `json:"description"`
-	Price       uint   `json:"price"`
+	ID          uint    `json:"id" gorm:"primaryKey"`
+	Description string  `json:"description"`
+	Price       float32 `json:"price"`
 }
